+++ internal/configprovider/manager_test.go
@@ -24,7 +24,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/config"
+	"go.opentelemetry.io/collector/config/configparser"
 	"go.opentelemetry.io/collector/config/experimental/configsource"
 	"go.uber.org/zap"
 )
@@ -64,7 +64,7 @@ func TestConfigSourceManager_NewManager(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			filename := path.Join("testdata", tt.file+".yaml")
-			parser, err := config.NewParserFromFile(filename)
+			parser, err := configparser.NewParserFromFile(filename)
 			require.NoError(t, err)
 
 			manager, err := NewManager(parser, zap.NewNop(), component.DefaultBuildInfo(), tt.factories)
@@ -102,7 +102,7 @@ func TestConfigSourceManager_Simple(t *testing.T) {
 		},
 	}
 
-	cp := config.NewParserFromStringMap(originalCfg)
+	cp := configparser.NewParserFromStringMap(originalCfg)
 
 	res, err := manager.Resolve(ctx, cp)
 	require.NoError(t, err)
@@ -135,7 +135,7 @@ func TestConfigSourceManager_ResolveRemoveConfigSourceSection(t *testing.T) {
 		"tstcfgsrc": &testConfigSource{},
 	})
 
-	res, err := manager.Resolve(context.Background(), config.NewParserFromStringMap(cfg))
+	res, err := manager.Resolve(context.Background(), configparser.NewParserFromStringMap(cfg))
 	require.NoError(t, err)
 	require.NotNil(t, res)
 
@@ -198,7 +198,7 @@ func TestConfigSourceManager_ResolveErrors(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			manager := newManager(tt.configSourceMap)
 
-			res, err := manager.Resolve(ctx, config.NewParserFromStringMap(tt.config))
+			res, err := manager.Resolve(ctx, configparser.NewParserFromStringMap(tt.config))
 			require.Error(t, err)
 			require.Nil(t, res)
 			require.NoError(t, manager.Close(ctx))
@@ -234,11 +234,11 @@ map:
 	})
 
 	file := path.Join("testdata", "yaml_injection.yaml")
-	cp, err := config.NewParserFromFile(file)
+	cp, err := configparser.NewParserFromFile(file)
 	require.NoError(t, err)
 
 	expectedFile := path.Join("testdata", "yaml_injection_expected.yaml")
-	expectedParser, err := config.NewParserFromFile(expectedFile)
+	expectedParser, err := configparser.NewParserFromFile(expectedFile)
 	require.NoError(t, err)
 	expectedCfg := expectedParser.ToStringMap()
 
@@ -263,11 +263,11 @@ func TestConfigSourceManager_ArraysAndMaps(t *testing.T) {
 	})
 
 	file := path.Join("testdata", "arrays_and_maps.yaml")
-	cp, err := config.NewParserFromFile(file)
+	cp, err := configparser.NewParserFromFile(file)
 	require.NoError(t, err)
 
 	expectedFile := path.Join("testdata", "arrays_and_maps_expected.yaml")
-	expectedParser, err := config.NewParserFromFile(expectedFile)
+	expectedParser, err := configparser.NewParserFromFile(expectedFile)
 	require.NoError(t, err)
 
 	res, err := manager.Resolve(ctx, cp)
@@ -313,11 +313,11 @@ func TestConfigSourceManager_ParamsHandling(t *testing.T) {
 	})
 
 	file := path.Join("testdata", "params_handling.yaml")
-	cp, err := config.NewParserFromFile(file)
+	cp, err := configparser.NewParserFromFile(file)
 	require.NoError(t, err)
 
 	expectedFile := path.Join("testdata", "params_handling_expected.yaml")
-	expectedParser, err := config.NewParserFromFile(expectedFile)
+	expectedParser, err := configparser.NewParserFromFile(expectedFile)
 	require.NoError(t, err)
 
 	res, err := manager.Resolve(ctx, cp)
@@ -349,7 +349,7 @@ func TestConfigSourceManager_WatchForUpdate(t *testing.T) {
 		},
 	}
 
-	cp := config.NewParserFromStringMap(originalCfg)
+	cp := configparser.NewParserFromStringMap(originalCfg)
 	_, err := manager.Resolve(ctx, cp)
 	require.NoError(t, err)
 
@@ -403,7 +403,7 @@ func TestConfigSourceManager_MultipleWatchForUpdate(t *testing.T) {
 		},
 	}
 
-	cp := config.NewParserFromStringMap(originalCfg)
+	cp := configparser.NewParserFromStringMap(originalCfg)
 	_, err := manager.Resolve(ctx, cp)
 	require.NoError(t, err)
 
@@ -452,11 +452,11 @@ func TestConfigSourceManager_EnvVarHandling(t *testing.T) {
 	})
 
 	file := path.Join("testdata", "envvar_cfgsrc_mix.yaml")
-	cp, err := config.NewParserFromFile(file)
+	cp, err := configparser.NewParserFromFile(file)
 	require.NoError(t, err)
 
 	expectedFile := path.Join("testdata", "envvar_cfgsrc_mix_expected.yaml")
-	expectedParser, err := config.NewParserFromFile(expectedFile)
+	expectedParser, err := configparser.NewParserFromFile(expectedFile)
 	require.NoError(t, err)
 
 	res, err := manager.Resolve(ctx, cp)