+++ internal/configprovider/parser.go
@@ -22,6 +22,7 @@ import (
 
 	"github.com/spf13/cast"
 	"go.opentelemetry.io/collector/config"
+	"go.opentelemetry.io/collector/config/configparser"
 	"go.opentelemetry.io/collector/config/experimental/configsource"
 )
 
@@ -39,7 +40,7 @@ type (
 
 // Load reads the configuration for ConfigSource objects from the given parser and returns a map
 // from the full name of config sources to the respective ConfigSettings.
-func Load(ctx context.Context, v *config.Parser, factories Factories) (map[string]ConfigSettings, error) {
+func Load(ctx context.Context, v *configparser.Parser, factories Factories) (map[string]ConfigSettings, error) {
 	processedParser, err := processParser(ctx, v)
 	if err != nil {
 		return nil, err
@@ -53,8 +54,8 @@ func Load(ctx context.Context, v *config.Parser, factories Factories) (map[strin
 	return cfgSrcSettings, nil
 }
 
-// processParser prepares a config.Parser to be used to load config source settings.
-func processParser(ctx context.Context, v *config.Parser) (*config.Parser, error) {
+// processParser prepares a configparser.Parser to be used to load config source settings.
+func processParser(ctx context.Context, v *configparser.Parser) (*configparser.Parser, error) {
 	// Use a manager to resolve environment variables with a syntax consistent with
 	// the config source usage.
 	manager := newManager(make(map[string]configsource.ConfigSource))
@@ -62,7 +63,7 @@ func processParser(ctx context.Context, v *config.Parser) (*config.Parser, error
 		_ = manager.Close(ctx)
 	}()
 
-	processedParser := config.NewParser()
+	processedParser := configparser.NewParser()
 	for _, key := range v.AllKeys() {
 		if !strings.HasPrefix(key, configSourcesKey) {
 			// In Load we only care about config sources, ignore everything else.
@@ -85,7 +86,7 @@ func loadSettings(css map[string]interface{}, factories Factories) (map[string]C
 
 	// Iterate over extensions and create a config for each.
 	for key, value := range css {
-		settingsParser := config.NewParserFromStringMap(cast.ToStringMap(value))
+		settingsParser := configparser.NewParserFromStringMap(cast.ToStringMap(value))
 
 		// Decode the key into type and fullName components.
 		componentID, err := config.NewIDFromString(key)