+++ internal/configprovider/parser_test.go
@@ -24,6 +24,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/config"
+	"go.opentelemetry.io/collector/config/configparser"
 	"go.opentelemetry.io/collector/config/experimental/configsource"
 )
 
@@ -118,7 +119,7 @@ func TestConfigSourceParser(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			cfgFile := path.Join("testdata", tt.file+".yaml")
-			v, err := config.NewParserFromFile(cfgFile)
+			v, err := configparser.NewParserFromFile(cfgFile)
 			require.NoError(t, err)
 
 			for key, value := range tt.envvars {