+++ internal/configsource/etcd2configsource/config_test.go
@@ -22,6 +22,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/config"
+	"go.opentelemetry.io/collector/config/configparser"
 	"go.uber.org/zap"
 
 	"github.com/signalfx/splunk-otel-collector/internal/configprovider"
@@ -29,7 +30,7 @@ import (
 
 func TestEtcd2LoadConfig(t *testing.T) {
 	fileName := path.Join("testdata", "config.yaml")
-	v, err := config.NewParserFromFile(fileName)
+	v, err := configparser.NewParserFromFile(fileName)
 	require.NoError(t, err)
 
 	factories := map[config.Type]configprovider.Factory{