+++ internal/receiver/smartagentreceiver/config_linux_test.go
@@ -38,8 +38,8 @@ func TestLoadConfigWithLinuxOnlyMonitors(t *testing.T) {
 
 	factory := NewFactory()
 	factories.Receivers[config.Type(typeStr)] = factory
-	cfg, err := configtest.LoadConfigFile(
-		t, path.Join(".", "testdata", "linux_config.yaml"), factories,
+	cfg, err := configtest.LoadConfig(
+		path.Join(".", "testdata", "linux_config.yaml"), factories,
 	)
 
 	require.NoError(t, err)